Format entire codebase using Black autoformatter

+++ arcade/experimental/examples/3d_cube.py
@@ -13,7 +13,6 @@ from arcade.gl import geometry
 
 
 class MyGame(arcade.Window):
-
     def __init__(self, width, height, title):
         super().__init__(width, height, title, resizable=True)
         self.set_vsync(True)
@@ -63,10 +62,12 @@ class MyGame(arcade.Window):
         self.clear()
         self.ctx.enable_only(self.ctx.CULL_FACE, self.ctx.DEPTH_TEST)
 
-        rotate = Matrix44.from_eulers((self.time, self.time * 0.77, self.time * 0.01), dtype='f4')
-        translate = Matrix44.from_translation((0, 0, -2.0), dtype='f4')
+        rotate = Matrix44.from_eulers(
+            (self.time, self.time * 0.77, self.time * 0.01), dtype="f4"
+        )
+        translate = Matrix44.from_translation((0, 0, -2.0), dtype="f4")
         modelview = translate * rotate
-        self.program['modelview'] = modelview.flatten()
+        self.program["modelview"] = modelview.flatten()
         self.cube.render(self.program)
 
     def on_update(self, dt):
@@ -77,7 +78,9 @@ class MyGame(arcade.Window):
         ratio = arcade.get_scaling_factor(self)
         self.ctx.viewport = 0, 0, int(width * ratio), int(height * ratio)
         aspect_ratio = width / height
-        self.program['projection'] = Matrix44.perspective_projection(60, aspect_ratio, 0.1, 100).flatten()
+        self.program["projection"] = Matrix44.perspective_projection(
+            60, aspect_ratio, 0.1, 100
+        ).flatten()
 
 
 if __name__ == "__main__":
